@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <RootNamespace>DotVast.HashTool.WinUI.Core</RootNamespace>
         <Platforms>x64;arm64;AnyCPU</Platforms>
         <ImplicitUsings>enable</ImplicitUsings>
